update i t

+++ src/amf/amf-sm.c
@@ -440,9 +440,9 @@ void amf_state_operational(ogs_fsm_t *s, amf_event_t *e)
             if (ECM_CONNECTED(amf_ue)) {
                /* Implcit NG release */
                 ogs_debug("Implicit NG release");
-                ogs_debug("    RAN_UE_NGAP_ID[%u] AMF_UE_NGAP_ID[%llu]",
+                ogs_debug("    RAN_UE_NGAP_ID[%d] AMF_UE_NGAP_ID[%lld]",
                       amf_ue->ran_ue->ran_ue_ngap_id,
-                      amf_ue->ran_ue->amf_ue_ngap_id);
+                      (long long)amf_ue->ran_ue->amf_ue_ngap_id);
                 ran_ue_remove(amf_ue->ran_ue);
             }
             amf_ue_associate_ran_ue(amf_ue, ran_ue);
