+++ src/amf/ngap-handler.c
@@ -316,9 +316,9 @@ void ngap_handle_initial_ue_message(amf_gnb_t *gnb, ogs_ngap_message_t *message)
                 if (ECM_CONNECTED(amf_ue)) {
                    /* Implcit NG release */
                     ogs_debug("Implicit NG release");
-                    ogs_debug("    RAN_UE_NGAP_ID[%u] AMF_UE_NGAP_ID[%lld]",
+                    ogs_debug("    RAN_UE_NGAP_ID[%d] AMF_UE_NGAP_ID[%lld]",
                           amf_ue->ran_ue->ran_ue_ngap_id,
-                          amf_ue->ran_ue->amf_ue_ngap_id);
+                          (long long)amf_ue->ran_ue->amf_ue_ngap_id);
                     ran_ue_remove(amf_ue->ran_ue);
                 }
                 amf_ue_associate_ran_ue(amf_ue, ran_ue);
@@ -340,10 +340,10 @@ void ngap_handle_initial_ue_message(amf_gnb_t *gnb, ogs_ngap_message_t *message)
         ogs_error("Not implemented UserLocationInformation[%d]",
                 UserLocationInformation->present);
 
-    ogs_debug("    RAN_UE_NGAP_ID[%u] AMF_UE_NGAP_ID[%llu] "
+    ogs_debug("    RAN_UE_NGAP_ID[%d] AMF_UE_NGAP_ID[%lld] "
             "TAC[%d] CellID[0x%llx]",
-        ran_ue->ran_ue_ngap_id, ran_ue->amf_ue_ngap_id,
-        ran_ue->saved.tai.tac.v, ran_ue->saved.cgi.cell_id);
+        ran_ue->ran_ue_ngap_id, (long long)ran_ue->amf_ue_ngap_id,
+        ran_ue->saved.tai.tac.v, (long long)ran_ue->saved.cgi.cell_id);
 
     ngap_send_to_nas(ran_ue,
             NGAP_ProcedureCode_id_InitialUEMessage, NAS_PDU);
